Final scripts to plot 421 paper plots
Large sections in collection scripts are skipped: smoothing, filtering, plotting. They may be used in future work.
The script for plotting collected info has renaming now for conditions and all plot outputs are better managed. Change the plot_output folder everywhere and tweak the axes.
The not smoothed and not filtered script was used for final data output for the paper.